spec(integration): extract shared fixture-bundle helper + set BUNDLE_FROZEN
Two cleanups on the rails_app fixture integration specs:

1. Extract the byte-identical bundle ladder + tracer-state scrub +
   cache reader from rails_app_spec.rb and narrow_ar_schema_spec.rb
   into spec/support/fixture_bundle_helper.rb. Each spec keeps its
   own per-spec constants (TRACER_ENV, MUTATION_MARKER) and
   subprocess runner; the shared FIXTURE_ROOT / CACHE_DIR /
   COVERAGE_DIR / REPORT_DIR / LOCAL_COVERAGE / SCRUB_DIRS constants
   and ensure_bundle_and_db / clear_tracer_state / load_cache_file
   methods now live in the helper. Per
   feedback_mutation_friendly_modules, helper uses `def self.x`
   over module_function for consistency with lib/ patterns.

2. Add BUNDLE_FROZEN=1 to the subprocess env passed by
   run_rspec_in_fixture in both specs. Bundler skips the per-`bundle
   exec` lockfile-resolve check when the lockfile is fully resolved
   ahead of time (which ensure_bundle_and_db guarantees at
   before(:all)). Saves ~0.5-1s per subprocess across the 24 cold +
   warm scenarios on the JRuby x Rails CI cell.

The brief described this as "hoist the bundle ladder to before(:all)
so the work runs once per spec-file load, not once per scenario" —
but the bundle ladder was already at before(:all). Reframed to the
DRY extraction + BUNDLE_FROZEN that actually deliver measurable
wins in the spirit of the lever.
